requirements.txt
streamlit
requests
pandas
numpy
pillow
openai
pyyaml
dotenv